requirements.txt
Flask==2.3.2
Werkzeug==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
mysqlclient==2.1.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
python-dotenv==1.0.0
